Add before handler in main

+++ main.go
@@ -4,10 +4,11 @@ import (
 	"chaostheory-task/internal/sqlitestore"
 	"database/sql"
 	"encoding/json"
-	"fmt"
+	"errors"
 	"log"
 	"net/http"
 	"os"
+	"strconv"
 
 	"github.com/go-chi/chi/v5"
 	"github.com/go-chi/chi/v5/middleware"
@@ -18,14 +19,12 @@ type ItemServer struct {
 }
 
 func NewItermServer(db *sql.DB) *ItemServer {
-	sqlitestore.NewItemStore(db)
 	return &ItemServer{
 		store: db,
 	}
 }
 
 func (is *ItemServer) ListHandler(w http.ResponseWriter, r *http.Request) {
-	// allItems := is.store.GetAllItems()
 	allItems := sqlitestore.GetAllItems(is.store)
 
 	js, err := json.Marshal(allItems)
@@ -55,15 +54,61 @@ func (is *ItemServer) AddHandler(w http.ResponseWriter, r *http.Request) {
 	w.Write(js)
 }
 
-func main() {
-	os.Remove("test.db")
-	log.Println("Creating ./data/test.db...")
-	file, err := os.Create("test.db")
+func (is *ItemServer) KeyHandler(w http.ResponseWriter, r *http.Request) {
+	key := chi.URLParam(r, "key")
+	item := sqlitestore.GetItemByKey(is.store, key)
+
+	js, err := json.Marshal(item)
 	if err != nil {
-		log.Fatal(err.Error())
+		http.Error(w, err.Error(), http.StatusInternalServerError)
+		return
 	}
-	file.Close()
-	db, _ := sql.Open("sqlite3", fmt.Sprintf("%s.db", "test"))
+	w.Header().Set("content-type", "application/json")
+	if item == nil {
+		http.Error(w, "Requested record does not exists", http.StatusNoContent)
+		return
+	}
+	w.Write(js)
+}
+
+func (is *ItemServer) DateBeforeHandler(w http.ResponseWriter, r *http.Request) {
+	y := chi.URLParam(r, "year")
+	m := chi.URLParam(r, "month")
+	d := chi.URLParam(r, "day")
+
+	year, err := strconv.Atoi(y)
+	if err != nil {
+		log.Fatal(err)
+		http.Error(w, "Expected integer year, month and day.", http.StatusBadRequest)
+	}
+	month, err := strconv.Atoi(m)
+	if err != nil {
+		log.Fatal(err)
+		http.Error(w, "Expected integer year, month and day.", http.StatusBadRequest)
+	}
+	day, err := strconv.Atoi(d)
+	if err != nil {
+		log.Fatal(err)
+		http.Error(w, "Expected integer year, month and day.", http.StatusBadRequest)
+	}
+
+}
+
+func main() {
+	if _, err := os.Stat("test.db"); errors.Is(err, os.ErrNotExist) {
+		// If .db file does not exist
+		log.Println("Creating test.db...")
+		file, err := os.Create("test.db")
+		if err != nil {
+			log.Fatal(err.Error())
+		}
+		file.Close()
+		db, _ := sql.Open("sqlite3", "test.db")
+		sqlitestore.NewDB(db)
+		db.Close()
+	}
+
+	db, _ := sql.Open("sqlite3", "test.db")
 	defer db.Close()
 	server := NewItermServer(db)
 
@@ -72,5 +117,7 @@ func main() {
 	r.Use(middleware.Logger)
 	r.Get("/list", server.ListHandler)
 	r.Post("/add", server.AddHandler)
+	r.Get("/key/{key}", server.KeyHandler)
+	r.Get("/date/before/{year}/{month}/{day}", server.DateBeforeHandler)
 	http.ListenAndServe(":3000", r)
 }
